requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djongo==1.3.6
dnspython==2.1.0
Faker==8.8.1
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
pymongo==3.11.4
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.5
psycopg2